Updated reqwest dependenciies
Added reqwest trace to the logger
Added ipfs tests tracing
