+++ git-remote-gosh/src/git_helper/test_utils.rs
@@ -1,30 +1,6 @@
-use crate::logger::set_log_verbosity;
 use git_object::bstr::ByteSlice;
 use git_repository::Repository;
-
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Once;
-use std::time::Duration;
 use std::{env, fs};
-use tokio::time::sleep;
-
-static TEST_LOGGER: Once = Once::new();
-static CALL_COUNT: AtomicUsize = AtomicUsize::new(0);
-
-pub async fn init_logger() {
-    TEST_LOGGER.call_once(|| {
-        set_log_verbosity(5);
-    });
-    CALL_COUNT.fetch_add(1, Ordering::SeqCst);
-}
-
-pub async fn shutdown_logger() {
-    CALL_COUNT.fetch_sub(1, Ordering::SeqCst);
-    if CALL_COUNT.load(Ordering::SeqCst) == 0 {
-        sleep(Duration::from_secs(10)).await;
-        // shutdown_tracer_provider();
-    }
-}
 
 pub fn setup_repo(name: &str, script_path: &str) -> anyhow::Result<Repository> {
     eprintln!("Testing push: current_dir = {:?}", env::current_dir()?);
